services/alert_sink_neo4j/requirements.txt
kafka-python==2.0.2
neo4j==5.19.0
python-dotenv==1.0.1
uuid